Move store order tracking to the only store class that needs it
Keeping track of file store records insertion order is actually quite slow. Moving order tracking to be just another stored key for the pending store makes a lot of sense, resulting in much faster performance for that store and other stores.

It also loosens the requirements of order tracking on other adapters in the future.